address warnings in win32 build (#550)

+++ src/ui/win32/bindings/GridBinding.cpp
@@ -917,7 +917,7 @@ LRESULT GridBinding::OnCustomDraw(NMLVCUSTOMDRAW* pCustomDraw)
                 break;
 
             case CDDS_ITEMPREPAINT | CDDS_SUBITEM:
-                if (pCustomDraw->iSubItem >= 0 && pCustomDraw->iSubItem < m_vColumns.size())
+                if (pCustomDraw->iSubItem >= 0 && pCustomDraw->iSubItem < gsl::narrow_cast<int>(m_vColumns.size()))
                 {
                     const auto& pColumn = *m_vColumns.at(pCustomDraw->iSubItem);
                     if (pColumn.GetTextColorProperty() != nullptr)
